Create Lab1_Problem1
Lab1_Problem1_ITS340
Problem #1: Write a program in Java to ask a user for input of a number. 
Check whether a number is even or odd and display a message to that effect. 
Keep asking for input from the user until the user enters a code that 
designates to QUIT.